Move UserDetailsFragment to MainActivity
