+++ presentation/src/androidTest/java/com/fernandocejas/android10/sample/test/view/fragment/UserlistFragmentTest.java
@@ -1,0 +1,7 @@
+package com.fernandocejas.android10.sample.test.view.fragment;
+
+/**
+ * Created by wolfgang on 28.09.15.
+ */
+public class UserlistFragmentTest {
+}
