+++ presentation/src/main/java/com/fernandocejas/android10/sample/presentation/navigation/Navigator.java
@@ -18,11 +18,10 @@ package com.fernandocejas.android10.sample.presentation.navigation;
 import android.app.Activity;
 import android.app.Fragment;
 import android.app.FragmentTransaction;
-import android.content.Context;
-import android.content.Intent;
 
 import com.fernandocejas.android10.sample.presentation.R;
-import com.fernandocejas.android10.sample.presentation.view.activity.UserDetailsActivity;
+import com.fernandocejas.android10.sample.presentation.view.fragment.LoadFragment;
+import com.fernandocejas.android10.sample.presentation.view.fragment.UserDetailsFragment;
 import com.fernandocejas.android10.sample.presentation.view.fragment.UserListFragment;
 
 import javax.inject.Inject;
@@ -42,7 +41,7 @@ public class Navigator {
   /**
    * Goes to the user list screen.
    *
-   * @param activity A Context needed to open the destiny activity.
+   * @param activity An activity needed to load the destination fragment.
    */
   public void navigateToUserList(Activity activity) {
     if (activity != null) {
@@ -53,16 +52,21 @@ public class Navigator {
   /**
    * Goes to the user details screen.
    *
-   * @param context A Context needed to open the destiny activity.
+   * @param activity An activity needed to open the destination fragment.
    */
-  public void navigateToUserDetails(Context context, int userId) {
-    if (context != null) {
-      Intent intentToLaunch = UserDetailsActivity.getCallingIntent(context, userId);
-      context.startActivity(intentToLaunch);
+  public void navigateToUserDetails(Activity activity, int userId) {
+    if (activity != null) {
+      replaceFragment(activity, R.id.container, UserDetailsFragment.newInstance(userId));
     }
   }
 
-    protected void replaceFragment(Activity activity, int containerViewId, Fragment fragment) {
+  public void navigateToLoadFragment(Activity activity) {
+    if (activity != null) {
+      replaceFragment(activity, R.id.container, LoadFragment.newInstance());
+    }
+  }
+
+  protected void replaceFragment(Activity activity, int containerViewId, Fragment fragment) {
     FragmentTransaction fragmentTransaction = activity.getFragmentManager().beginTransaction();
     fragmentTransaction.replace(containerViewId, fragment);
     fragmentTransaction.commit();
