--- presentation/src/main/res/layout/activity_user_details.xml
@@ -1,12 +1,0 @@
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    >
-
-  <FrameLayout
-      android:id="@+id/fl_fragment"
-      android:layout_width="match_parent"
-      android:layout_height="match_parent"
-      />
-
-</LinearLayout>
